Add "uid-range" and "systemd-cgroup" system features
"uid-range" provides 65536 UIDs to a build and runs the build as root
in its user namespace. "systemd-cgroup" allows the build to mount the
systemd cgroup controller (needed for running systemd-nspawn and NixOS
containers).

Also, add a configuration option "auto-allocate-uids" which is needed
to enable these features, and some experimental feature gates.

So to enable support for containers you need the following in
nix.conf:

  experimental-features = auto-allocate-uids systemd-cgroup
  auto-allocate-uids = true
  system-features = uid-range systemd-cgroup

+++ src/libstore/cgroup.cc
@@ -4,6 +4,7 @@
 #include "util.hh"
 
 #include <chrono>
+#include <unordered_set>
 
 #include <dirent.h>
 
@@ -19,7 +20,7 @@ std::map<std::string, std::string> getCgroups(const Path & cgroupFile)
         if (!std::regex_match(line, match, regex))
             throw Error("invalid line '%s' in '%s'", line, cgroupFile);
 
-        std::string name = hasPrefix(match[2], "name=") ? std::string(match[2], 5) : match[2];
+        std::string name = hasPrefix(std::string(match[2]), "name=") ? std::string(match[2], 5) : match[2];
         cgroups.insert_or_assign(name, match[3]);
     }
 
@@ -28,12 +29,16 @@ std::map<std::string, std::string> getCgroups(const Path & cgroupFile)
 
 void destroyCgroup(const Path & cgroup)
 {
+    if (!pathExists(cgroup)) return;
+
     for (auto & entry : readDirectory(cgroup)) {
         if (entry.type != DT_DIR) continue;
         destroyCgroup(cgroup + "/" + entry.name);
     }
 
     int round = 1;
+
+    std::unordered_set<pid_t> pidsShown;
 
     while (true) {
         auto pids = tokenizeString<std::vector<std::string>>(readFile(cgroup + "/cgroup.procs"));
@@ -46,13 +51,23 @@ void destroyCgroup(const Path & cgroup)
         for (auto & pid_s : pids) {
             pid_t pid;
             if (!string2Int(pid_s, pid)) throw Error("invalid pid '%s'", pid);
+            if (pidsShown.insert(pid).second) {
+                try {
+                    auto cmdline = readFile(fmt("/proc/%d/cmdline", pid));
+                    using namespace std::string_literals;
+                    warn("killing stray builder process %d (%s)...",
+                        pid, trim(replaceStrings(cmdline, "\0"s, " ")));
+                } catch (SysError &) {
+                }
+            }
             // FIXME: pid wraparound
             if (kill(pid, SIGKILL) == -1 && errno != ESRCH)
                 throw SysError("killing member %d of cgroup '%s'", pid, cgroup);
         }
 
         auto sleep = std::chrono::milliseconds((int) std::pow(2.0, std::min(round, 10)));
-        printError("waiting for %d ms for cgroup '%s' to become empty", sleep.count(), cgroup);
+        if (sleep.count() > 100)
+            printError("waiting for %d ms for cgroup '%s' to become empty", sleep.count(), cgroup);
         std::this_thread::sleep_for(sleep);
         round++;
     }
